pyproject.toml
[tool.poetry]
name = "griptape_statemachine"
version = "0.1.0"
description = ""
authors = ["Collin Dutter <collin@griptape.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
griptape = "^0.32"
python-statemachine = {extras = ["diagrams"], version = "^2.3.6"}
pyyaml = "^6.0.2"
schema = "^0.7.7"


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.380"
ruff = "^0.6.4"
typos = "^1.25.0"
pre-commit = "^3.8.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
   "ALL"
]
ignore = [
   "D",
   "COM812", # missing-trailing-comma -- See https://github.com/astral-sh/ruff/issues/9216 
   "ANN003",
   'T201',
   "TD",
   "FIX",
   "E501"
]
[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = true

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"attr".msg = "The attr module is deprecated, use attrs instead."

[tool.pyright]
venvPath = "."
venv = ".venv"
include = [
   "griptape_statemachine"
]
exclude = [
    "**/__pycache__",
]
pythonVersion = "3.11"
enableExperimentalFeatures = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
